REimagine_Waste_backend/requirements.txt
absl-py==0.14.1
adal==1.2.7
altgraph==0.17
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.2.10
attrs==21.2.0
autopep8==1.4.4
autopy==4.0.0
awsebcli==3.18.1
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==4.1.0
BlurWindow==1.2.1
botocore==1.15.49
cachetools==4.2.2
cement==2.8.2
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cryptography==35.0.0
ctypes-callable==1.0.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
docutils==0.15.2
EB==0.1.5
entrypoints==0.3
et-xmlfile==1.1.0
exponent-server-sdk==2.0.0
filelock==3.0.12
future==0.16.0
googletrans==2.4.0
graphviz==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.7
importlib-metadata==4.8.2
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mediapipe==0.8.7.3
mistune==0.8.4
MouseInfo==0.1.3
msrest==0.6.21
msrestazure==0.6.4
mysql-connector-python==8.0.24
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
notebook-as-pdf==0.5.0
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==3.0.7
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.5.9
pefile==2021.5.24
phonenumbers==8.12.7
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2020.5.28
playsound==1.2.2
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.18.0
psycopg2==2.9.2
py2exe==0.10.4.0
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycparser==2.20
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
pyppeteer==0.2.6
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.27
python-dateutil==2.8.2
pyttsx3==2.88
PyTweening==1.0.3
pytz==2020.1
pywhatkit==4.7
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==1.1.4
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.20.1
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
semantic-version==2.5.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.7
virtualenv==20.0.21
virtualenv-clone==0.5.4
wcwidth==0.1.9
webencodings==0.5.1
websockets==9.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
yarn==0.0.8
zipp==3.6.0
